[SPARK-36915][INFRA] Pin actions to a full length commit SHA
### What changes were proposed in this pull request?
Pinning github actions to a SHA

### Why are the changes needed?
Pinning an action to a full length commit SHA is currently the only way to use an action as
an immutable release. Pinning to a particular SHA helps mitigate the risk of a bad actor adding a
backdoor to the action's repository, as they would need to generate a SHA-1 collision for
a valid Git object payload.

https://docs.github.com/en/actions/security-guides/security-hardening-for-github-actions#using-third-party-actions

https://github.com/ossf/scorecard/blob/main/docs/checks.md#pinned-dependencies

### Does this PR introduce _any_ user-facing change?
Running github action and checking the SHA with the existing repository

### How was this patch tested?
Running the GitHub action

Closes #34163 from naveensrinivasan/naveen/feat/pin-github-actions.

Lead-authored-by: Hyukjin Kwon <gurwls223@gmail.com>
Co-authored-by: naveen <172697+naveensrinivasan@users.noreply.github.com>
Signed-off-by: Sean Owen <srowen@gmail.com>